+++ index.js
@@ -1,5 +1,0 @@
-/*!
- * Copyright (c) 2019-2021 Digital Bazaar, Inc. All rights reserved.
- */
-
-export {default as ProfileManager} from './ProfileManager.js';