+++ test/web/10-api.js
@@ -1,8 +1,8 @@
 /*!
  * Copyright (c) 2019-2022 Digital Bazaar, Inc. All rights reserved.
  */
-import {ProfileManager} from 'bedrock-web-profile-manager';
-import {ProfileService} from 'bedrock-web-profile';
+import {ProfileManager} from '@bedrock/web-profile-manager';
+import {ProfileService} from '@bedrock/web-profile';
 import sinon from 'sinon';
 import {mockData} from './mock.data.js';