+++ test/web/20-collection.js
@@ -1,7 +1,7 @@
 /*!
  * Copyright (c) 2019-2022 Digital Bazaar, Inc. All rights reserved.
  */
-import {ProfileManager} from 'bedrock-web-profile-manager';
+import {ProfileManager} from '@bedrock/web-profile-manager';
 import {EdvClient} from '@digitalbazaar/edv-client';
 
 const ACCOUNT_ID = 'urn:uuid:ffaf5d84-7dc2-4f7b-9825-cc8d2e5a5d06';